Integracion vuelo paquete turistico, hotel + api responses y exceptions

+++ src/main/java/com/viajaYa/viajaYa/services/PaqueteTuristicoService.java
@@ -1,8 +1,15 @@
 package com.viajaYa.viajaYa.services;
 
+import com.viajaYa.viajaYa.models.dtos.PaqueteTuristicoDTO;
+import com.viajaYa.viajaYa.models.HotelModel;
 import com.viajaYa.viajaYa.models.PaqueteTuristicoModel;
+import com.viajaYa.viajaYa.models.VueloModel;
 import com.viajaYa.viajaYa.repositories.IPaqueteTuristicoRepository;
+import com.viajaYa.viajaYa.services.interfaces.IHotelService;
 import com.viajaYa.viajaYa.services.interfaces.IPaqueteTuristicoService;
+import com.viajaYa.viajaYa.services.interfaces.IVueloService;
+import com.viajaYa.viajaYa.services.mappers.IMapper;
+import com.viajaYa.viajaYa.utils.exceptions.BusinessException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -14,46 +21,68 @@ public class PaqueteTuristicoService implements IPaqueteTuristicoService {
 
     @Autowired
     IPaqueteTuristicoRepository paqueteTuristicoRepository;
+    @Autowired
+    IHotelService hotelService;
+    @Autowired
+    IVueloService vueloService;
+    @Autowired
+    IMapper<PaqueteTuristicoDTO, PaqueteTuristicoModel> mapper;
 
     @Override
     public ArrayList<PaqueteTuristicoModel> getPaquetes(){
-        return (ArrayList<PaqueteTuristicoModel>) paqueteTuristicoRepository.findAll();
+         return (ArrayList<PaqueteTuristicoModel>) paqueteTuristicoRepository.findAll();
     }
 
     @Override
-    public Optional<PaqueteTuristicoModel> getByid(int id){
-        return paqueteTuristicoRepository.findById(id);
-    }
-
-    @Override
-    public PaqueteTuristicoModel savePaqueteTuristico(PaqueteTuristicoModel paquete){
-        return paqueteTuristicoRepository.save(paquete);
-    }
-
-    @Override
-    public PaqueteTuristicoModel updateById(PaqueteTuristicoModel request, int id){
-        PaqueteTuristicoModel paquete = paqueteTuristicoRepository.findById(id).get();
-        paquete.setNombrePaquete(request.getNombrePaquete());
-        paquete.setDestino(request.getDestino());
-        paquete.setPrecio(request.getPrecio());
-        paquete.setServiciosIncluidos(request.getServiciosIncluidos());
-        paquete.setFechaSalida(request.getFechaSalida());
-        paquete.setIdVuelo(request.getIdVuelo());
-        paquete.setIdHotel(request.getIdHotel());
-
-        paqueteTuristicoRepository.save(paquete);
-
+    public Optional<PaqueteTuristicoModel> getByid(Long id){
+        Optional<PaqueteTuristicoModel> paquete = paqueteTuristicoRepository.findById(id);
+        if(paquete.isEmpty()){
+            throw new BusinessException("Paquete turistico con id " + id + " no encontrado");
+        }
         return paquete;
     }
 
     @Override
-    public boolean deletePaquete(int id){
-        try{
-            paqueteTuristicoRepository.deleteById(id);
-            return true;
+    public PaqueteTuristicoModel savePaqueteTuristico(PaqueteTuristicoDTO dto){
+
+        PaqueteTuristicoModel paquete = mapper.toEntity(dto);
+        return paqueteTuristicoRepository.save(paquete);
+    }
+
+    @Override
+    public PaqueteTuristicoModel updateById(PaqueteTuristicoDTO request, Long id){
+        PaqueteTuristicoModel paqueteExistente = paqueteTuristicoRepository.findById(id)
+                .orElseThrow(() -> new BusinessException("Paquete turístico con id " + id + " no encontrado"));
+
+        PaqueteTuristicoModel paqueteDTO = mapper.toEntity(request);
+
+        paqueteExistente.setNombrePaquete(paqueteDTO.getNombrePaquete());
+        paqueteExistente.setDestino(paqueteDTO.getDestino());
+        paqueteExistente.setPrecio(paqueteDTO.getPrecio());
+        paqueteExistente.setServiciosIncluidos(paqueteDTO.getServiciosIncluidos());
+        paqueteExistente.setFechaSalida(paqueteDTO.getFechaSalida());
+
+        if (request.getIdVuelo() != null) {
+            VueloModel vuelo = vueloService.getVueloById(request.getIdVuelo())
+                    .orElseThrow(() -> new BusinessException("Vuelo no encontrado"));
+            paqueteExistente.setVuelo(vuelo);
         }
-        catch (Exception e){
-            return false;
+        if (request.getIdHotel() != null) {
+            HotelModel hotel = hotelService.getHotelId(request.getIdHotel())
+                    .orElseThrow(() -> new BusinessException("Hotel no encontrado"));
+            paqueteExistente.setHotel(hotel);
         }
+
+        return paqueteTuristicoRepository.save(paqueteExistente);
+    }
+
+    @Override
+    public boolean deletePaquete(Long id){
+        Optional<PaqueteTuristicoModel> paquete = paqueteTuristicoRepository.findById(id);
+        if(paquete.isEmpty()){
+            throw new BusinessException("Paquete turistico con id " + id + " no encontrado");
+        }
+        paqueteTuristicoRepository.deleteById(id);
+        return true;
     }
 }
